+++ packages/core/src/contexts/ScaffoldContext.ts
@@ -6,12 +6,18 @@
 import Context from './Context';
 import { Arguments } from '../types';
 
-export default class ScaffoldContext extends Context {
+export interface ScaffoldArgs {
+  action: string;
+  generator: string;
+  dry: boolean;
+}
+
+export default class ScaffoldContext<T = ScaffoldArgs> extends Context<T> {
   action: string;
 
   generator: string;
 
-  constructor(args: Arguments, generator: string, action: string) {
+  constructor(args: Arguments<T>, generator: string, action: string) {
     super(args);
 
     this.generator = generator;