+++ packages/driver-babel/src/BabelDriver.ts
@@ -5,8 +5,8 @@
 
 import path from 'path';
 import rimraf from 'rimraf';
-import { Driver, DriverContext } from '@beemo/core';
-import { BabelConfig } from './types';
+import { Driver, DriverArgs, DriverContext } from '@beemo/core';
+import { BabelArgs, BabelConfig } from './types';
 
 // Success: Writes file list to stdout
 // Failure: Throws SyntaxError to stderr
@@ -18,6 +18,7 @@ export default class BabelDriver extends Driver<BabelConfig> {
       configOption: '--config-file',
       description: this.tool.msg('app:babelDescription'),
       title: 'Babel',
+      watchOptions: ['-w', '--watch'],
     });
 
     this.setCommandOptions({
@@ -42,7 +43,7 @@ export default class BabelDriver extends Driver<BabelConfig> {
   /**
    * Automatically clean the target folder if --out-dir is used.
    */
-  handleCleanTarget = ({ args }: DriverContext) => {
+  handleCleanTarget = ({ args }: DriverContext<DriverArgs & BabelArgs & { clean?: boolean }>) => {
     if (args.clean && args.outDir) {
       rimraf.sync(path.resolve(args.outDir));
     }