+++ packages/core/CHANGELOG.md
@@ -7,6 +7,12 @@
   - `parse()` was renamed to `args()`.
   - `run(options, tool)` was renamed to `execute(context, options)`.
 
+#### 🚀 New
+
+- Added _experimental_ support for capturing watch output from drivers.
+- Added `watchOptions` metadata field to `Driver`. Defines CLI watch options to intercept for the
+  defined driver.
+
 # 0.6.0
 
 #### 💥 Breaking